+++ axelor-stock/src/main/java/com/axelor/apps/stock/service/StockMoveLineServiceImpl.java
@@ -64,6 +64,7 @@ public class StockMoveLineServiceImpl implements StockMoveLineService {
   protected StockMoveService stockMoveService;
   private TrackingNumberService trackingNumberService;
   protected StockMoveLineRepository stockMoveLineRepository;
+  protected StockLocationLineService stockLocationLineService;
 
   @Inject
   public StockMoveLineServiceImpl(
@@ -71,12 +72,14 @@ public class StockMoveLineServiceImpl implements StockMoveLineService {
       AppBaseService appBaseService,
       AppStockService appStockService,
       StockMoveService stockMoveService,
-      StockMoveLineRepository stockMoveLineRepository) {
+      StockMoveLineRepository stockMoveLineRepository,
+      StockLocationLineService stockLocationLineService) {
     this.trackingNumberService = trackingNumberService;
     this.appBaseService = appBaseService;
     this.appStockService = appStockService;
     this.stockMoveService = stockMoveService;
     this.stockMoveLineRepository = stockMoveLineRepository;
+    this.stockLocationLineService = stockLocationLineService;
   }
 
   /**
@@ -888,5 +891,46 @@ public class StockMoveLineServiceImpl implements StockMoveLineService {
       stockMoveLineRepository.save(newStockMoveLine);
     }
     stockMoveLineRepository.remove(stockMoveLine);
+  }
+
+  @Override
+  public void updateAvailableQty(StockMoveLine stockMoveLine, StockLocation stockLocation) {
+    BigDecimal availableQty = BigDecimal.ZERO;
+    BigDecimal availableQtyForProduct = BigDecimal.ZERO;
+
+    if (stockMoveLine.getProduct() != null) {
+      if (stockMoveLine.getProduct().getTrackingNumberConfiguration() != null) {
+
+        if (stockMoveLine.getTrackingNumber() != null) {
+          StockLocationLine stockLocationLine =
+              stockLocationLineService.getDetailLocationLine(
+                  stockLocation, stockMoveLine.getProduct(), stockMoveLine.getTrackingNumber());
+
+          if (stockLocationLine != null) {
+            availableQty = stockLocationLine.getCurrentQty();
+          }
+        }
+
+        if (availableQty.compareTo(stockMoveLine.getRealQty()) < 0) {
+          StockLocationLine stockLocationLineForProduct =
+              stockLocationLineService.getStockLocationLine(
+                  stockLocation, stockMoveLine.getProduct());
+
+          if (stockLocationLineForProduct != null) {
+            availableQtyForProduct = stockLocationLineForProduct.getCurrentQty();
+          }
+        }
+      } else {
+        StockLocationLine stockLocationLine =
+            stockLocationLineService.getStockLocationLine(
+                stockLocation, stockMoveLine.getProduct());
+
+        if (stockLocationLine != null) {
+          availableQty = stockLocationLine.getCurrentQty();
+        }
+      }
+    }
+    stockMoveLine.setAvailableQty(availableQty);
+    stockMoveLine.setAvailableQtyForProduct(availableQtyForProduct);
   }
 }