+++ CHANGELOG.md
@@ -68,6 +68,7 @@
 - Manufacturing order sequence is now configured in company config, and can be configured by workshop.
 - Studio: Add description field in workflow
 - Group menu assistant: Add support for roles management.  
+- Stock Move : add three transient fields 'availableQty','availableQtyForProduct' and 'availableStatus' to know the availability of the product on stock move line.
 
 ## Improvements
 - Move project resource menu entry to project configuration menu.